+++ lib/commands/migrate_databse_seeder_command.dart
@@ -12,7 +12,7 @@ class MigrateDatabaseSeederCommand implements Command {
 
   @override
   void execute(List<String> arguments) async {
-    print('\x1B[32m Database seed started \x1B[0m');
+    stdout.writeln('\x1B[32m Database seed started \x1B[0m');
     Process process = await Process.start(
       'dart',
       [
@@ -25,7 +25,7 @@ class MigrateDatabaseSeederCommand implements Command {
       List lines = data.split("\n");
       for (String line in lines) {
         if (line.isNotEmpty) {
-          print(line);
+          stdout.writeln(line);
         }
       }
     }