+++ lib/commands/serve_command.dart
@@ -109,7 +109,7 @@ class ServeCommand implements Command {
       List lines = data.split("\n");
       for (String line in lines) {
         if (line.isNotEmpty) {
-          print(line);
+          stdout.writeln(line);
         }
       }
     });
@@ -118,7 +118,7 @@ class ServeCommand implements Command {
       List lines = data.split("\n");
       for (String line in lines) {
         if (line.isNotEmpty) {
-          print(line);
+          stdout.writeln(line);
         }
       }
     });