+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.2.0
+
+- Add DB Seed
+- Add Migrate fresh
+- Fix migration name
+- Refactor `print` to `stdout.writeln`
+
 ## 1.1.0
 
 - Add Terminate open port